requirements.txt
pdf2image
paddleocr
paddlepaddle
fastapi
uvicorn
python-multipart
gradio
opencv-python
numpy
Pillow